+++ examples/bar/package.json
@@ -31,7 +31,7 @@
     "@cycle/core": "^6.0.3",
     "@cycle/dom": "^9.4.0",
     "d3": "^3.5.16",
-    "grizzy": "0.1.2",
+    "grizzy": "0.1.3",
     "rx": "^4.1.0"
   }
 }
