--- npm-debug.log
@@ -1,48 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/Cellar/node/6.6.0/bin/node',
-1 verbose cli   '/usr/local/bin/npm',
-1 verbose cli   'run',
-1 verbose cli   'lint' ]
-2 info using npm@3.10.7
-3 info using node@v6.6.0
-4 verbose run-script [ 'prelint', 'lint', 'postlint' ]
-5 info lifecycle grizzy@0.1.2~prelint: grizzy@0.1.2
-6 silly lifecycle grizzy@0.1.2~prelint: no script for prelint, continuing
-7 info lifecycle grizzy@0.1.2~lint: grizzy@0.1.2
-8 verbose lifecycle grizzy@0.1.2~lint: unsafe-perm in lifecycle true
-9 verbose lifecycle grizzy@0.1.2~lint: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/jameshaley/apps-js/d3-exp/grizzy/node_modules/.bin:/usr/local/Cellar/node/6.6.0/bin:/usr/local/bin:/Users/jameshaley/.rvm/gems/ruby-2.2.0/bin:/Users/jameshaley/.rvm/gems/ruby-2.2.0@global/bin:/Users/jameshaley/.rvm/rubies/ruby-2.2.0/bin:/Library/Frameworks/Python.framework/Versions/3.4/bin:/Library/Frameworks/Python.framework/Versions/2.7/bin:/usr/bin:/bin:/usr/sbin:/sbin:/usr/local/bin:/opt/X11/bin:/usr/local/smlnj/bin:/Users/jameshaley/.rvm/bin:/Users/jameshaley/.bin
-10 verbose lifecycle grizzy@0.1.2~lint: CWD: /Users/jameshaley/apps-js/d3-exp/grizzy
-11 silly lifecycle grizzy@0.1.2~lint: Args: [ '-c', 'eslint src' ]
-12 silly lifecycle grizzy@0.1.2~lint: Returned: code: 1  signal: null
-13 info lifecycle grizzy@0.1.2~lint: Failed to exec lint script
-14 verbose stack Error: grizzy@0.1.2 lint: `eslint src`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:255:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:191:7)
-14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:40:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:191:7)
-14 verbose stack     at maybeClose (internal/child_process.js:877:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
-15 verbose pkgid grizzy@0.1.2
-16 verbose cwd /Users/jameshaley/apps-js/d3-exp/grizzy
-17 error Darwin 15.6.0
-18 error argv "/usr/local/Cellar/node/6.6.0/bin/node" "/usr/local/bin/npm" "run" "lint"
-19 error node v6.6.0
-20 error npm  v3.10.7
-21 error code ELIFECYCLE
-22 error grizzy@0.1.2 lint: `eslint src`
-22 error Exit status 1
-23 error Failed at the grizzy@0.1.2 lint script 'eslint src'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the grizzy package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     eslint src
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs grizzy
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls grizzy
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
